rename `projection.rs` to `temporal_axes.rs`
